pyproject.toml
[tool.poetry]
name = "foundry-alexa"
version = "0.1.0"
description = ""
authors = ["Liam Hughes"]

[tool.poetry.dependencies]
python = "^3.10"
ask-sdk-core = "^1.17.1"
ask-sdk-model = "^1.34.2"
boto3 = "^1.24.60"
aws-cdk-lib = "^2.39.0"
"aws-cdk.aws-lambda-python-alpha" = "^2.39.0-alpha.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
